% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_exptilt_ipsi.R
\name{LF_exptilt_ipsi}
\alias{LF_exptilt_ipsi}
\title{Exponential Tilt Likelihood Factor for Incremental Propensity Scores}
\format{\code{R6::R6Class} object.}
\value{
\code{tmle3::LF_base} object
}
\description{
Exponential Tilt Likelihood Factor for Incremental Propensity Scores

Exponential Tilt Likelihood Factor for Incremental Propensity Scores
}
\section{Constructor}{

  \code{define_lf(LF_exptilt_ipsi, name, type = "density", likelihood_base,
    shift_param, treatment_task, control_task, ...)}

  \describe{
    \item{\code{name}}{character, the name of the factor. Should match a node
          name in the nodes specified by \code{tmle3::tmle3_Task$npsem}.
    }
    \item{\code{likelihood_base}}{The trained \code{tmle3::Likelihood}
          object, for use in generating a re-scaled likelihood factor.
    }
    \item{\code{shift_param}}{\code{numeric}, specification of the magnitude
          of the desired shift (a multiplier for the propensity score).
    }
    \item{\code{treatment_task}}{\code{tmle3::tmle3_Task} object created from
          setting the intervention to the treatment condition: do(A = 1).
    }
    \item{\code{control_task}}{\code{tmle3::tmle3_Task} object created from
          setting the intervention to the control condition: do(A = 0).
    }
    \item{\code{...}}{Not currently used.
    }
  }
}

\section{Fields}{

\describe{
    \item{\code{likelihood_base}}{The trained \code{tmle3::Likelihood}
          object, for use in generating a re-scaled likelihood factor.
    }
    \item{\code{shift_param}}{\code{numeric}, specification of the magnitude
          of the desired shift (a multiplier for the propensity score).
    }
    \item{\code{treatment_task}}{\code{tmle3::tmle3_Task} object created from
          setting the intervention to the treatment condition: do(A = 1).
    }
    \item{\code{control_task}}{\code{tmle3::tmle3_Task} object created from
          setting the intervention to the control condition: do(A = 0).
    }
    \item{\code{...}}{Additional arguments passed to the base class.
    }
  }
}

\references{
\describe{
  \item{"Nonparametric Causal Effects Based on Incremental Propensity Score
        Interventions."}{Edward's JASA paper...}
  \item{"Causal Mediation Analysis for Stochastic Interventions"}{Díaz, Iván
        and Hejazi, Nima S (2019). submitted.}
}
}
\concept{Likelihood objects}
\keyword{data}
\section{Super class}{
\code{\link[tmle3:LF_base]{tmle3::LF_base}} -> \code{LF_exptilt_ipsi}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LF_exptilt_ipsi$new()}}
\item \href{#method-get_mean}{\code{LF_exptilt_ipsi$get_mean()}}
\item \href{#method-get_density}{\code{LF_exptilt_ipsi$get_density()}}
\item \href{#method-cf_values}{\code{LF_exptilt_ipsi$cf_values()}}
\item \href{#method-clone}{\code{LF_exptilt_ipsi$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="LF_base" data-id="delayed_train">}\href{../../tmle3/html/LF_base.html#method-delayed_train}{\code{tmle3::LF_base$delayed_train()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="LF_base" data-id="get_likelihood">}\href{../../tmle3/html/LF_base.html#method-get_likelihood}{\code{tmle3::LF_base$get_likelihood()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="LF_base" data-id="print">}\href{../../tmle3/html/LF_base.html#method-print}{\code{tmle3::LF_base$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="LF_base" data-id="train">}\href{../../tmle3/html/LF_base.html#method-train}{\code{tmle3::LF_base$train()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_exptilt_ipsi$new(
  name,
  likelihood_base,
  shift_param,
  treatment_task,
  control_task,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_mean"></a>}}
\subsection{Method \code{get_mean()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_exptilt_ipsi$get_mean(tmle_task, fold_number)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_density"></a>}}
\subsection{Method \code{get_density()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_exptilt_ipsi$get_density(tmle_task, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cf_values"></a>}}
\subsection{Method \code{cf_values()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_exptilt_ipsi$cf_values(tmle_task)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_exptilt_ipsi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
