% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_NIE.R
\name{make_NIE_psi_Z_task}
\alias{make_NIE_psi_Z_task}
\alias{make_NIE_psi_Z0_task}
\alias{make_NIE_psi_Z1_task}
\title{Make task for derived likelihood factor psi_Z(W) for NIE}
\usage{
make_NIE_psi_Z0_task(tmle_task, likelihood)

make_NIE_psi_Z1_task(tmle_task, likelihood)
}
\arguments{
\item{tmle_task}{A \code{[tmle3]{tmle3_Task}} specifying the data and NPSEM
for use in constructing components required for TML estimation.}

\item{likelihood}{A trained \code{[tmle3]{Likelihood}}, constructed via the
\code{\link{stochastic_mediation_likelihood}} helper.}
}
\description{
Make task for derived likelihood factor psi_Z(W) for NIE
}
\keyword{internal}
