% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mediator_task_fun_factory}
\alias{mediator_task_fun_factory}
\title{Factory function that returns an \code{[sl3]{sl3_Task}} generator function
used in fitting the appropriate mediator derived conditional likelihood
factor for NDE/NIE/medshift.}
\usage{
mediator_task_fun_factory(
  param_type = c("NDE", "NIE", "medshift"),
  subset_value = NA
)
}
\arguments{
\item{param_type}{A \code{character}, giving the target parameter type for
which to construct the psi_Z \code{[tmle3]{tmle3_Task}}.}

\item{subset_value}{A \code{numeric}, which determines the treatment value
used to subset the \code{[tmle3]{tmle3_Task}}. \code{NA} by default for no
subsetting.}
}
\value{
A \code{function} that takes a \code{[tmle3]{tmle3_Task}} and a
\code{[tmle3]{Likelihood}} and returns a new \code{[sl3]{sl3_Task}}.
}
\description{
Factory function that returns an \code{[sl3]{sl3_Task}} generator function
used in fitting the appropriate mediator derived conditional likelihood
factor for NDE/NIE/medshift.
}
\keyword{internal}
