% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_stochastic_mediation.R
\name{Param_medshift}
\alias{Param_medshift}
\title{Parameter for mediation effect under stochastic interventions}
\format{\code{\link{R6Class}} object.}
\value{
\code{Param_base} object
}
\description{
Parameter definition class. See https://arxiv.org/abs/1901.02776
}
\section{Constructor}{

  \code{define_param(Param_medshift, shift_param, ..., outcome_node)}

  \describe{
    \item{\code{observed_likelihood}}{A \code{\link{Likelihood}}
          corresponding to the observed likelihood.
    }
    \item{\code{shift_param}}{\code{numeric}, specification of the magnitude
          of the desired shift (a multiplier for the propensity score).
    }
    \item{\code{...}}{Not currently used.
    }
    \item{\code{outcome_node}}{character, the name of the node that should be
          treated as the outcome
    }
  }
}

\section{Fields}{

\describe{
    \item{\code{cf_likelihood}}{the counterfactual likelihood under the
          stochastic intervention on mediators and treatment.
    }
    \item{\code{lf_exptilt}}{Object derived from \code{\link{LF_base}} for
          assessing the stochastic mediation intervention.
    }
    \item{\code{treatment_task}}{\code{\link{tmle3_Task}} object created from
          setting the intervention to the treatment condition: do(A = 1).
    }
    \item{\code{control_task}}{\code{\link{tmle3_Task}} object created from
          setting the intervention to the control condition: do(A = 0).
    }
    \item{\code{shift_param}}{\code{numeric}, specification of the magnitude
          of the desired shift (a multiplier for the propensity score).
    }
}
}

\concept{Parameters}
\keyword{data}
\section{Super class}{
\code{\link[tmle3:Param_base]{tmle3::Param_base}} -> \code{Param_medshift}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Param_medshift$new()}}
\item \href{#method-clever_covariates}{\code{Param_medshift$clever_covariates()}}
\item \href{#method-estimates}{\code{Param_medshift$estimates()}}
\item \href{#method-clone}{\code{Param_medshift$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="Param_base" data-id="print">}\href{../../tmle3/html/Param_base.html#method-print}{\code{tmle3::Param_base$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_medshift$new(observed_likelihood, shift_param, ..., outcome_node = "Y")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clever_covariates"></a>}}
\subsection{Method \code{clever_covariates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_medshift$clever_covariates(tmle_task = NULL, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimates"></a>}}
\subsection{Method \code{estimates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_medshift$estimates(tmle_task = NULL, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_medshift$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
