% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_NIE.R
\name{Param_NIE}
\alias{Param_NIE}
\title{Parameter for the natural indirect effect}
\format{\code{\link{R6Class}} object.}
\value{
\code{Param_base} object
}
\description{
Parameter definition class. See
\url{https://www.ncbi.nlm.nih.gov/pubmed/22499725}
}
\section{Constructor}{

\code{define_param(Param_ATT, observed_likelihood, intervention_list, ...,
                     outcome_node)}

\describe{
\item{\code{observed_likelihood}}{A \code{\link[tmle3]{Likelihood}}
corresponding to the observed likelihood.}
\item{\code{...}}{Not currently used.}
\item{\code{outcome_node}}{A \code{character}, giving the name of the
node that should be treated as the outcome.}
}
}

\section{Fields}{

\describe{
\item{\code{cf_likelihood_treatment}}{The counterfactual likelihood for
the treatment.}
\item{\code{cf_likelihood_control}}{The counterfactual likelihood for
the control.}
\item{\code{treatment_task}}{\code{\link[tmle3]{tmle3_Task}} created by
setting the intervention to the treatment condition: do(A = 1).}
\item{\code{control_task}}{\code{\link[tmle3]{tmle3_Task}} created by
setting the intervention to the control condition: do(A = 0).}
}
}

\seealso{
Other Parameters: 
\code{\link{Param_NDE}},
\code{\link{Param_medshift}}
}
\concept{Parameters}
\keyword{data}
