% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{psi_Z_task_factory}
\alias{psi_Z_task_factory}
\title{Factory function that returns an \code{[sl3]{sl3_Task}} generator function
used in fitting the psi_Z(W) derived likelihood factor for NDE/NIE.}
\usage{
psi_Z_task_factory(subset_value, param_type = c("NDE", "NIE"))
}
\arguments{
\item{subset_value}{A \code{numeric}, which determines the treatment value used to
subset the \code{[tmle3]{tmle3_Task}}.}

\item{param_type}{A \code{character}, giving the target parameter type for
which to construct the psi_Z \code{[tmle3]{tmle3_Task}}.}

\item{tmle_task}{A \code{[tmle3]{tmle3_Task}} specifying the data and NPSEM
for use in constructing components required for TML estimation.}

\item{likelihood}{A trained \code{[tmle3]{Likelihood}}, constructed via the
\code{\link{stochastic_mediation_likelihood}} helper.}
}
\value{
A \code{function} that takes a \code{[tmle3]{tmle3_Task}} and a
\code{[tmle3]{Likelihood}} and returns a new \code{[sl3]{sl3_Task}}.
}
\description{
Factory function that returns an \code{[sl3]{sl3_Task}} generator function
used in fitting the psi_Z(W) derived likelihood factor for NDE/NIE.
}
\keyword{internal}
