% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_NDE.R
\name{tmle3_Spec_NDE}
\alias{tmle3_Spec_NDE}
\title{Defines a TML Estimator for Outcome under Joint Static Intervention on
Treatment and Mediator}
\description{
Defines a TML Estimator for Outcome under Joint Static Intervention on
Treatment and Mediator

Defines a TML Estimator for Outcome under Joint Static Intervention on
Treatment and Mediator
}
\section{Super class}{
\code{\link[tmle3:tmle3_Spec]{tmle3::tmle3_Spec}} -> \code{tmle3_Spec_NDE}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{tmle3_Spec_NDE$new()}}
\item \href{#method-make_tmle_task}{\code{tmle3_Spec_NDE$make_tmle_task()}}
\item \href{#method-make_initial_likelihood}{\code{tmle3_Spec_NDE$make_initial_likelihood()}}
\item \href{#method-make_params}{\code{tmle3_Spec_NDE$make_params()}}
\item \href{#method-make_updater}{\code{tmle3_Spec_NDE$make_updater()}}
\item \href{#method-clone}{\code{tmle3_Spec_NDE$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="tmle3_Spec" data-id="make_targeted_likelihood">}\href{../../tmle3/html/tmle3_Spec.html#method-make_targeted_likelihood}{\code{tmle3::tmle3_Spec$make_targeted_likelihood()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_NDE$new(
  e_learners,
  psi_Z_learners,
  max_iter = 10000,
  step_size = 1e-06,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_tmle_task"></a>}}
\subsection{Method \code{make_tmle_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_NDE$make_tmle_task(data, node_list, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_initial_likelihood"></a>}}
\subsection{Method \code{make_initial_likelihood()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_NDE$make_initial_likelihood(tmle_task, learner_list = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_params"></a>}}
\subsection{Method \code{make_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_NDE$make_params(tmle_task, targeted_likelihood)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_updater"></a>}}
\subsection{Method \code{make_updater()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_NDE$make_updater()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tmle3_Spec_NDE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
