% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_NIE.R
\name{tmle_NIE}
\alias{tmle_NIE}
\title{TML Estimator for the Natural Indirect Effect}
\usage{
tmle_NIE(e_learners, psi_Z_learners, max_iter = 10000, step_size = 1e-06, ...)
}
\arguments{
\item{e_learners}{A \code{\link[sl3]{Stack}} (or other learner class that
inherits from \code{\link[sl3]{Lrnr_base}}), containing a single or set of
instantiated learners from \pkg{sl3}, to be used in fitting a cleverly
parameterized propensity score that conditions on the mediators, i.e.,
\eqn{e = P(A \mid Z, W)}.}

\item{psi_Z_learners}{A \code{\link[sl3]{Stack}} (or other learner class
that inherits from \code{\link[sl3]{Lrnr_base}}), containing a single or
set of instantiated learners from \pkg{sl3}, to be used in a regression of
a pseudo-outcome on the baseline covariates, i.e.,
\eqn{psi_Z(W) = E[m(A = 1, Z, W) - m(A = 0, Z, W) \mid A = 0, W]}.}

\item{max_iter}{A \code{numeric} setting the maximum iterations allowed in
the targeting step based on universal least favorable submodels.}

\item{step_size}{A \code{numeric} giving the step size (\code{delta_epsilon}
in \code{\link[tmle3]{tmle3}}) to be used in the targeting step based on
universal least favorable submodels.}

\item{...}{Additional arguments (currently unused).}
}
\description{
O = (W, A, Z, Y)
W = Covariates (possibly multivariate)
A = Treatment (binary or categorical)
Z = Mediators (binary or categorical; possibly multivariate)
Y = Outcome (binary or bounded continuous)
}
