% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_NDE.R
\name{make_NDE_psi_Z_task}
\alias{make_NDE_psi_Z_task}
\title{Make task for derived likelihood factor psi_Z(W) for NDE}
\usage{
make_NDE_psi_Z_task(tmle_task, likelihood)
}
\arguments{
\item{tmle_task}{A \code{tmle3_Task} object specifying the data and the
NPSEM for use in constructing elements of TML estimator.}

\item{likelihood}{A trained \code{Likelihood} object from \code{tmle3},
constructed via the helper function \code{stochastic_mediation_likelihood}.}
}
\description{
Make task for derived likelihood factor psi_Z(W) for NDE
}
\keyword{internal}
