% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_NIE.R
\name{tmle_NIE}
\alias{tmle_NIE}
\title{Natural Indirect Effect}
\usage{
tmle_NIE(treatment_level, control_level)
}
\arguments{
\item{treatment_level}{the level of A that corresponds to treatment}

\item{control_level}{the level of A that corresponds to a control or reference level}
}
\description{
O=(W,A,Z,Y)
W=Covariates
A=Treatment (binary or categorical)
Z=Mediator
Y=Outcome (binary or bounded continuous)
}
